- also fix math for planting site - refactor service - use non-raw so we get string formatted correctly
