+++ src/components/PlantingSites/PlantingSitesList.tsx
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useRef, useState } from 'react';
 import { Box, CircularProgress, Grid, Typography } from '@mui/material';
 import { Button, theme } from '@terraware/web-components';
 import { useDeviceInfo } from '@terraware/web-components/utils';
-import { SearchService } from 'src/services';
+import { TrackingService } from 'src/services';
 import { SearchResponseElement, SearchSortOrder } from 'src/types/Search';
 import strings from 'src/strings';
 import useDebounce from 'src/utils/useDebounce';
@@ -41,7 +41,7 @@ export default function PlantingSitesList(): JSX.Element {
           ],
         }
       : undefined;
-    const apiSearchResults = await SearchService.searchPlantingSites(
+    const apiSearchResults = await TrackingService.searchPlantingSites(
       selectedOrganization.id,
       searchField,
       searchSortOrder
